create types file for querybuilder and rework column props
